+++ ludwig/benchmarking/utils.py
@@ -1,12 +1,18 @@
-import logging
 import os
+import asyncio
+import functools
+import logging
+
 from types import ModuleType
-from typing import Any, Dict, Union
+from typing import Any, Dict, List, Tuple, Union
+
+from concurrent.futures import ThreadPoolExecutor
 
 import fsspec
 import shutil
 import pandas as pd
 
+from ludwig.utils.data_utils import load_yaml
 from ludwig.constants import CATEGORY
 from ludwig.datasets.base_dataset import BaseDataset
 from ludwig.globals import CONFIG_YAML
@@ -16,7 +22,7 @@ from ludwig.utils.fs_utils import get_fs_and_path
 
 
 def load_from_module(
-    dataset_module: Union[BaseDataset, ModuleType], output_feature: Dict[str, str], subsample_frac: float = 1
+        dataset_module: Union[BaseDataset, ModuleType], output_feature: Dict[str, str], subsample_frac: float = 1
 ) -> pd.DataFrame:
     """Load the ludwig dataset, optionally subsamples it, and returns a repeatable split. A stratified split is
     used for classification datasets.
@@ -65,15 +71,75 @@ def export_artifacts(experiment: Dict[str, str], experiment_output_directory: st
         )
 
 
+def download_artifacts(
+        bench_config_path: str, base_experiment: str, experimental_experiment: str, download_base_path: str,
+        local_dir: str = "benchmarking_summaries",
+) -> Tuple[str, List[str]]:
+    """Download benchmarking artifacts for two experiments.
+
+    bench_config: bench config file. Can be the same one that was used to run
+        these experiments.
+    base_experiment: name of the experiment we're comparing against.
+    experimental_experiment: name of the experiment we're comparing.
+    download_base_path: base path under which live the stored artifacts of
+        the benchmarking experiments.
+    """
+    bench_config = load_yaml(bench_config_path)
+    protocol, _ = fsspec.core.split_protocol(download_base_path)
+    fs, _ = get_fs_and_path(download_base_path)
+    os.makedirs(local_dir, exist_ok=True)
+
+    coroutines = []
+    for experiment in bench_config["datasets"]:
+        dataset_name = experiment["dataset_name"]
+        for experiment_name in [base_experiment, experimental_experiment]:
+            coroutines.append(download_one(fs, download_base_path, dataset_name, experiment_name, local_dir))
+    loop = asyncio.get_event_loop()
+    futures = asyncio.gather(*coroutines, return_exceptions=True)
+    downloaded_names = loop.run_until_complete(futures)
+
+    dataset_names = [experiment_tuple[0] for experiment_tuple in set(downloaded_names) if experiment_tuple[0]]
+    assert len(set(experiment_tuple[1] for experiment_tuple in downloaded_names)) == 1 and downloaded_names[0][
+        1] == local_dir, "Experiments not downloaded to the same path"
+
+    return local_dir, dataset_names
+
+
+async def download_one(
+        fs, download_base_path: str, dataset_name: str, experiment_name: str, local_dir: str
+) -> Tuple[str, str]:
+    """Download `config.yaml` and `report.json` for an experiment.
+
+    fs: filesystem to use to download.
+    download_base_path: base path under which live the stored artifacts of
+        the benchmarking experiments.
+    dataset_name: name of the dataset we ran the experiments on.
+    experiment_name: name of the experiment (e.g. `v0.5.3_with_bert`)
+    local_dir: local directory under which the artifacts will be downloaded.
+    """
+    loop = asyncio.get_running_loop()
+    local_experiment_dir = os.path.join(local_dir, dataset_name, experiment_name)
+    remote_experiment_directory = os.path.join(download_base_path, dataset_name, experiment_name)
+    os.makedirs(local_experiment_dir, exist_ok=True)
+    try:
+        with ThreadPoolExecutor() as pool:
+            func = functools.partial(
+                fs.get,
+                remote_experiment_directory,
+                local_experiment_dir,
+                recursive=True,
+            )
+            await loop.run_in_executor(pool, func)
+    except:
+        logging.exception(f"Couldn't download experiment *{experiment_name}* of dataset *{dataset_name}*.")
+        return "", local_dir
+    return dataset_name, local_dir
+
+
 def delete_model_checkpoints(output_directory: str):
     shutil.rmtree(os.path.join(output_directory, "model", "training_checkpoints"), ignore_errors=True)
     if os.path.isfile(os.path.join(output_directory, "model", "model_weights")):
         os.remove(os.path.join(output_directory, "model", "model_weights"))
-
-
-def flatten_dict(d: Dict[str, Any], sep: str = ".") -> Dict[str, Any]:
-    [flat_dict] = pd.json_normalize(d, sep=sep).to_dict(orient="records")
-    return flat_dict
 
 
 def format_time(time_us):
